Rename _2daObject to AuroraTable and fix namespaces
