+++ AuroraIOTests/Source/AuroraTableTests.cs
@@ -10,14 +10,12 @@ using System.Reflection;
 
 namespace AuroraIOTests.Source {
     [TestClass]
-    public class _2DAObjectTests {
-
-        _2DACoder coder = new _2DACoder();
+    public class AuroraTableTests {
 
         [TestMethod]
         public void testWrite()
         {
-            var array = _2daObjectStubs.stub1();
+            var array = AuroraTableStubs.stub1();
 
             Snapshot.Verify(array);
         }
@@ -25,7 +23,7 @@ namespace AuroraIOTests.Source {
         [TestMethod]
         public void testAddRow()
         {
-            var array = _2daObjectStubs.stub1();
+            var array = AuroraTableStubs.stub1();
             var row = array.addRow();
 
             Snapshot.Verify(array);
@@ -33,7 +31,7 @@ namespace AuroraIOTests.Source {
 
         [TestMethod]
         public void testAddRowValues() {
-            var array = _2daObjectStubs.stub1();
+            var array = AuroraTableStubs.stub1();
             array.addRow(new string[] { "c1r4", "c2r4", "c3r4", "c4r4" });
 
             Snapshot.Verify(array);
@@ -42,7 +40,7 @@ namespace AuroraIOTests.Source {
         [TestMethod]
         public void testAddRowMap()
         {
-            var array = _2daObjectStubs.stub1();
+            var array = AuroraTableStubs.stub1();
             array.addRow(new Dictionary<string, string>
             {
                 { "column0", "c0r4" },
@@ -57,7 +55,7 @@ namespace AuroraIOTests.Source {
         [TestMethod]
         public void testModifyValueInNewRowWithIndex()
         {
-            var array = _2daObjectStubs.stub1();
+            var array = AuroraTableStubs.stub1();
             var row = array.addRow();
             row[0] = "mod";
 
@@ -67,7 +65,7 @@ namespace AuroraIOTests.Source {
         [TestMethod]
         public void testModifyValueInNewRowWithKey()
         {
-            var array = _2daObjectStubs.stub1();
+            var array = AuroraTableStubs.stub1();
             var row = array.addRow();
             row["column0"] = "mod";
 
@@ -77,7 +75,7 @@ namespace AuroraIOTests.Source {
         [TestMethod]
         public void testModifyValueInExistingRowWithIndex()
         {
-            var array = _2daObjectStubs.stub1();
+            var array = AuroraTableStubs.stub1();
             array[0][0] = "mod";
 
             Snapshot.Verify(array);
@@ -86,7 +84,7 @@ namespace AuroraIOTests.Source {
         [TestMethod]
         public void testModifyValueInExistingRowWithKey()
         {
-            var array = _2daObjectStubs.stub1();
+            var array = AuroraTableStubs.stub1();
             array[0]["column0"] = "mod";
 
             Snapshot.Verify(array);
@@ -95,7 +93,7 @@ namespace AuroraIOTests.Source {
         [TestMethod]
         public void testModifyMultipleValuesInExistingRow()
         {
-            var array = _2daObjectStubs.stub1();
+            var array = AuroraTableStubs.stub1();
             array[0]["column0"] = "mod";
             array[2]["column3"] = "mod";
 
@@ -105,7 +103,7 @@ namespace AuroraIOTests.Source {
         [TestMethod]
         public void testClearValue()
         {
-            var array = _2daObjectStubs.stub1();
+            var array = AuroraTableStubs.stub1();
             array[0]["column0"] = null;
 
             Snapshot.Verify(array);
