+++ AuroraIOTests/Source/Stubs/AuroraTableStubs.cs
@@ -1,15 +1,16 @@
 ﻿using AuroraIO.Source.Models._2da;
+using AuroraIO.Source.Models.Table;
 using System;
 using System.Collections.Generic;
 using System.Text;
 
 namespace AuroraIOTests.Source.Stubs
 {
-    public static class _2daObjectStubs
+    public static class AuroraTableStubs
     {
-        public static _2DAObject stub1()
+        public static AuroraTable stub1()
         {
-            return new _2DAObject(
+            return new AuroraTable(
                 new string[] { "column0", "column1", "column2",	"column3" },
                 new string[][] {
                     new string[] { "c0r0", "c1r0", "c2r0", "c3r0" },
