--- AuroraIOTests/Source/testEncode4ComplexField_actual
@@ -1,14 +1,0 @@
-type: gff
-fields:
-  field0:
-    type: cexostring
-    value: string0
-  field1:
-    type: cexostring
-    value: string1
-  field2:
-    type: cexostring
-    value: string2
-  field3:
-    type: cexostring
-    value: string3
