--- AuroraIOTests/Source/testEncode4SimpleField_actual
@@ -1,14 +1,0 @@
-type: gff
-fields:
-  field0:
-    type: byte
-    value: 0
-  field1:
-    type: byte
-    value: 1
-  field2:
-    type: byte
-    value: 2
-  field3:
-    type: byte
-    value: 3
