--- AuroraIOTests/Source/testEncodeCExoLocString_actual
@@ -1,6 +1,0 @@
-type: gff
-fields:
-  value:
-    type: cexolocstring
-    value:
-      strref: -1
