--- AuroraIOTests/Source/testEncodeDouble_actual
@@ -1,5 +1,0 @@
-type: gff
-fields:
-  value:
-    type: double
-    value: 100
