--- AuroraIOTests/Source/testEncodeEmpty_actual
@@ -1,2 +1,0 @@
-type: gff
-fields:
