--- AuroraIOTests/Source/testEncodeFloat_actual
@@ -1,5 +1,0 @@
-type: gff
-fields:
-  value:
-    type: float
-    value: 100
