--- AuroraIOTests/Source/testEncodeList_actual
@@ -1,11 +1,0 @@
-type: gff
-fields:
-  value:
-    type: list
-    value:
-      - struct:
-        id: 0
-        fields:
-          value:
-            type: int
-            value: 100
