--- AuroraIOTests/Source/testEncodeMixedListStructField_actual
@@ -1,36 +1,0 @@
-type: gff
-fields:
-  field0:
-    type: list
-    value:
-      - struct:
-        id: 0
-        fields:
-          field0-item:
-            type: byte
-            value: 0
-  field1:
-    type: list
-    value:
-      - struct:
-        id: 1
-        fields:
-          field1-item:
-            type: byte
-            value: 1
-  field2:
-    type: struct
-    value:
-      id: 2
-      fields:
-        field2-item:
-          type: byte
-          value: 2
-  field3:
-    type: struct
-    value:
-      id: 3
-      fields:
-        field3-item:
-          type: byte
-          value: 3
