--- AuroraIOTests/Source/testEncodeMixedSimpleComplexField_actual
@@ -1,14 +1,0 @@
-type: gff
-fields:
-  field0:
-    type: int
-    value: 0
-  field1:
-    type: int
-    value: 1
-  field2:
-    type: cexostring
-    value: string2
-  field3:
-    type: cexostring
-    value: string3
