--- AuroraIOTests/Source/testEncodeNestedListField_actual
@@ -1,19 +1,0 @@
-type: gff
-fields:
-  field0:
-    type: list
-    value:
-      - struct:
-        id: 0
-        fields:
-          field1:
-            type: struct
-            value:
-              id: 1
-              fields:
-                field2:
-                  type: int
-                  value: 100
-  field3:
-    type: int
-    value: 100
