--- AuroraIOTests/Source/testEncodeNestedStructField_actual
@@ -1,13 +1,0 @@
-type: gff
-fields:
-  field0:
-    type: struct
-    value:
-      id: 0
-      fields:
-        field1:
-          type: int
-          value: 100
-  field2:
-    type: int
-    value: 100
