--- AuroraIOTests/Source/testEncodeQuaternion_actual
@@ -1,9 +1,0 @@
-type: gff
-fields:
-  value:
-    type: quaternion
-    value:
-      w: 0.1
-      x: 0.2
-      y: 0.3
-      z: 0.4
