--- AuroraIOTests/Source/testEncodeShort_actual
@@ -1,5 +1,0 @@
-type: gff
-fields:
-  value:
-    type: short
-    value: 100
