--- AuroraIOTests/Source/testEncodeVector_actual
@@ -1,8 +1,0 @@
-type: gff
-fields:
-  value:
-    type: vector
-    value:
-      x: 0.1
-      y: 0.2
-      z: 0.3
