--- AuroraIOTests/Source/testEncodeWord_actual
@@ -1,5 +1,0 @@
-type: gff
-fields:
-  value:
-    type: word
-    value: 100
