--- AuroraIOTests/Source/testModifyValueInExistingRowWithKey_actual
@@ -1,5 +1,0 @@
-column0	column1	column2	column3
-mod	c1r0	c2r0	c3r0
-c0r1	c1r1	c2r1	c3r1
-c0r2	c1r2	c2r2	c3r2
-c0r3	c1r3	c2r3	c3r3
