--- AuroraIOTests/Source/testNestedStructDictionary_actual
@@ -1,24 +1,0 @@
-type: gff
-fields:
-  struct:
-    type: struct
-    value:
-      id: -1
-      fields:
-        int:
-          type: int
-          value: 0
-        float:
-          type: float
-          value: 0
-        struct:
-          type: struct
-          value:
-            id: -1
-            fields:
-              int:
-                type: int
-                value: 0
-              float:
-                type: float
-                value: 0
