--- AuroraIOTests/Source/testPopulateDictionary_actual
@@ -1,103 +1,0 @@
-type: gff
-fields:
-  field_byte:
-    type: byte
-    value: 0
-  field_char:
-    type: char
-    value: c
-  field_word:
-    type: word
-    value: 0
-  field_short:
-    type: short
-    value: 0
-  field_dword:
-    type: dword
-    value: 0
-  field_int:
-    type: int
-    value: 0
-  field_dword64:
-    type: dword64
-    value: 0
-  field_int64:
-    type: int64
-    value: 0
-  field_float:
-    type: float
-    value: 0
-  field_double:
-    type: double
-    value: 0
-  field_cexoString:
-    type: cexostring
-    value: cexostring
-  field_resref:
-    type: resref
-    value: resref
-  field_cexolocstring:
-    type: cexolocstring
-    value:
-      strref: -1
-      englishfemale: female
-  void:
-    type: void
-    value: 000102030405060708090A0B0C0D0E0F
-  field_struct:
-    type: struct
-    value:
-      id: -1
-      fields:
-        field_struct.int:
-          type: int
-          value: 0
-        field_struct.byte:
-          type: byte
-          value: 0
-  field_list:
-    type: list
-    value:
-      - struct:
-        id: 0
-        fields:
-          sub1:
-            type: int
-            value: 0
-          sub2:
-            type: cexostring
-            value: testString
-      - struct:
-        id: 1
-        fields:
-          sub1:
-            type: int
-            value: 0
-          sub2:
-            type: cexostring
-            value: testString
-      - struct:
-        id: 2
-        fields:
-          sub1:
-            type: int
-            value: 0
-          sub2:
-            type: cexostring
-            value: testString
-  field_quaternion:
-    type: quaternion
-    value:
-      w: 0
-      x: 0
-      y: 0
-      z: 0
-  field_vector:
-    type: vector
-    value:
-      x: 0
-      y: 0
-      z: 0
-  field_strref:
-    type: strref
-    value: 0
